kgqa_docker/kgqa/requirements.txt
PyYAML
sentence_transformers==2.4.0
yaspin
tqdm
pandas==2.1.4
openai==1.17.0
tabulate
pandas
termcolor
rdflib
graphviz
flask
flask-cors
streamlit
gravis
faiss-cpu
requests
snowflake-connector-python[pandas]
yaspin
psycopg2-binary